+++ packages/big-design/src/components/MultiSelect/MultiSelect.tsx
@@ -3,6 +3,7 @@ import React, {
   cloneElement,
   createRef,
   isValidElement,
+  LabelHTMLAttributes,
   RefObject,
   useCallback,
   useEffect,
@@ -319,8 +320,11 @@ export const MultiSelect = typedMemo(
         );
       }
 
-      if (isValidElement(label) && label.type === FormControlLabel) {
-        return cloneElement<React.LabelHTMLAttributes<HTMLLabelElement>>(label, getLabelProps());
+      if (
+        isValidElement<LabelHTMLAttributes<HTMLLabelElement>>(label) &&
+        label.type === FormControlLabel
+      ) {
+        return cloneElement(label, getLabelProps());
       }
 
       warning('label must be either a string or a FormControlLabel component.');