+++ packages/big-design/src/components/Textarea/Textarea.tsx
@@ -1,4 +1,11 @@
-import React, { cloneElement, forwardRef, isValidElement, Ref, useMemo } from 'react';
+import React, {
+  cloneElement,
+  forwardRef,
+  isValidElement,
+  LabelHTMLAttributes,
+  Ref,
+  useMemo,
+} from 'react';
 
 import { useUniqueId } from '../../hooks';
 import { typedMemo, warning } from '../../utils';
@@ -51,8 +58,11 @@ const StyleableTextarea: React.FC<TextareaProps & PrivateProps> = ({
       );
     }
 
-    if (isValidElement(label) && label.type === FormControlLabel) {
-      return cloneElement<React.LabelHTMLAttributes<HTMLLabelElement>>(label, {
+    if (
+      isValidElement<LabelHTMLAttributes<HTMLLabelElement>>(label) &&
+      label.type === FormControlLabel
+    ) {
+      return cloneElement(label, {
         id: labelId,
         htmlFor: id,
       });